Fixed cloning of tournaments.  Corrected how various tournament dates are calculated.  Confirmed deletion of event from tournament.

+++ client/src/app/tournament/tournament-config/tournament-event-config-list/tournament-event-config-list-container.component.ts
@@ -5,6 +5,7 @@ import {TournamentEventConfigService} from '../tournament-event-config.service';
 import {ConfirmationPopupComponent} from '../../../shared/confirmation-popup/confirmation-popup.component';
 import {MatDialog} from '@angular/material/dialog';
 import {TournamentEventConfigListComponent} from './tournament-event-config-list.component';
+import {tap} from 'rxjs/operators';
 
 @Component({
   selector: 'app-tournament-event-config-list-container',
@@ -41,6 +42,8 @@ export class TournamentEventConfigListContainerComponent implements OnInit, OnCh
 
   events$: Observable<TournamentEvent []>;
 
+  private eventsWithEntries: number [] = [];
+
   @ViewChild(TournamentEventConfigListComponent)
   tournamentEventConfigListComponent: TournamentEventConfigListComponent;
 
@@ -55,19 +58,32 @@ export class TournamentEventConfigListContainerComponent implements OnInit, OnCh
   ngOnChanges(changes: SimpleChanges) {
     if (changes.tournamentId != null) {
       const tournamentId = changes.tournamentId.currentValue;
-      this.tournamentEventConfigService.loadTournamentEvents(tournamentId);
+      this.tournamentEventConfigService.loadTournamentEvents(tournamentId)
+        .pipe(tap((events: TournamentEvent[]) => {
+          const eventsWithEntries: number [] = [];
+          for (const event of events) {
+            if (event.numEntries > 0) {
+              eventsWithEntries.push(event.id);
+            }
+          }
+          this.eventsWithEntries = eventsWithEntries;
+      }));
     }
   }
 
   onDelete(eventId: number) {
+    const hasEntries = (this.eventsWithEntries.indexOf(eventId) >= 0);
+    const message = (hasEntries)
+      ? 'Warning: There are entries into this event.  You must first remove all entries from this event. Press \'OK\' to close'
+      : 'Are you sure you want to delete this event.  Press \'OK\' to proceed';
     const config = {
       width: '450px', height: '250px', data: {
-        message: 'There are entries into this event.  Are you sure you want to delete this event.  Press \'OK\' to proceed',
+        message: message, showOk: !hasEntries
       }
     };
     const dialogRef = this.dialog.open(ConfirmationPopupComponent, config);
     dialogRef.afterClosed().subscribe(result => {
-      if (result === 'ok') {
+      if (result === 'ok' && !hasEntries) {
         this.tournamentEventConfigService.delete(eventId);
       }
     });
